requirements.txt
africastalking==1.2.4
alabaster==0.7.12
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.5.2
atomicwrites==1.3.0
attrs==20.3.0
autopep8==1.5.4
backcall==0.1.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto==2.49.0
boto3==1.18.31
botocore==1.21.31
celery==4.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==36.0.1
decorator==4.0.9
defusedxml==0.7.1
dispatch==1.3.16
distlib==0.3.2
dj-database-url==0.5.0
Django==4.1.2
django-allauth==0.49.0
django-appconf==1.0.4
django-bootstrap==0.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.0
django-celery==3.3.1
django-celery-beat==1.6.0
django-ckeditor==6.0.0
django-ckeditor-5==0.0.13
django-cors-headers==3.8.0
django-countries==7.3
django-cprofile-middleware==1.0.5
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-etc==1.3.0
django-extensions==3.1.2
django-filter==2.2.0
django-heroku==0.0.0
django-imagekit==4.0.2
django-jquery==3.1.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mathfilters==1.0.0
django-phonenumbers==1.0.1
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-redis==4.12.1
django-sendgrid-v5==0.9.1
django-ses==3.0.1
django-simple-captcha==0.5.17
django-sslserver==0.22
django-static-jquery==2.1.4
django-storages==1.11.1
django-timezone-field==3.1
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.21.4
environ==1.0
environment==1.0.0
extension==1.0.0
Faker==0.9.1
fanstatic==1.1
fcm-django==0.3.7
filelock==3.0.12
Flask==1.1.2
flower==0.9.4
fontawesome==5.10.1.post1
future==0.18.2
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httplib2==0.9.2
humanize==0.5.1
hyperframe==5.2.0
idna==3.3
importlib-metadata==3.3.0
inflection==0.5.1
iniconfig==1.1.1
ipython-genutils==0.1.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.0.1
jupyter-core==4.4.0
kombu==4.6.11
Markdown==3.3.4
markdown2==2.4.5
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.1.0
num2words==0.5.10
oauthlib==3.2.0
p5py==1.0.0
packaging==21.3
parso==0.4.0
paystack==1.5.0
paystackapi==2.0.0
pexpect==4.0.1
phonenumbers==8.12.45
phonenumberslite==7.3.1
pickleshare==0.7.2
pilkit==2.0
Pillow==9.0.1
Pillow-PIL==0.1.dev0
pluggy==0.13.1
prompt-toolkit==3.0.2
psycopg2-binary==2.8.6
ptyprocess==0.5.1
py==1.10.0
pycodestyle==2.6.0
pycparser==2.21
pyfcm==1.4.9
Pygments==2.5.2
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.15.1
pytest==6.2.0
pytest-cov==2.10.1
pytest-django==3.9.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.13.0
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2021.3
pyuca==1.2
redis==3.5.3
requests==2.27.1
requests-async==0.6.2
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
schema==0.7.2
sendgrid==6.4.8
sentry-sdk==0.14.3
shutilwhich==1.1.0
simplegeneric==0.8.1
simplejson==3.17.2
six==1.15.0
slugify==0.0.1
smtpapi==0.3.1
soupsieve==2.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.0
text-unidecode==1.2
toml==0.10.2
tornado==6.1
traitlets==4.2.1
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==1.5.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.8
vine==1.3.0
virtualenv==20.4.7
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.6
Werkzeug==1.0.1
whitenoise==5.3.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0